# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5e2456320cfc6233ae9336d9
  title: Manage_Film_Example
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /filmmaker/:
    post:
      x-skaffolder-id: 5e2456330cfc6233ae933717
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5e2456330cfc6233ae9336f9
      x-skaffolder-resource: FilmMaker
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - FilmMaker
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmMaker"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                surname:
                  type: string
    get:
      x-skaffolder-id: 5e2456340cfc6233ae93371c
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5e2456330cfc6233ae9336f9
      x-skaffolder-resource: FilmMaker
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - FilmMaker
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FilmMaker"
  /filmmaker/{id}:
    delete:
      x-skaffolder-id: 5e2456340cfc6233ae93371d
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e2456330cfc6233ae9336f9
      x-skaffolder-resource: FilmMaker
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - FilmMaker
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e2456330cfc6233ae933714
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e2456330cfc6233ae9336f9
      x-skaffolder-resource: FilmMaker
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - FilmMaker
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmMaker"
    post:
      x-skaffolder-id: 5e2456330cfc6233ae933716
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5e2456330cfc6233ae9336f9
      x-skaffolder-resource: FilmMaker
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - FilmMaker
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmMaker"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilmMaker"
    
  #-----  User  -----
...
  /user/:
    post:
      x-skaffolder-id: 5e2456320cfc6233ae9336e9
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5e2456320cfc6233ae9336e4
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5e2456320cfc6233ae9336eb
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5e2456320cfc6233ae9336e4
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /user/{id}:
    delete:
      x-skaffolder-id: 5e2456320cfc6233ae9336e8
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e2456320cfc6233ae9336e4
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e2456320cfc6233ae9336ec
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e2456320cfc6233ae9336e4
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5e2456320cfc6233ae9336ed
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5e2456320cfc6233ae9336e4
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
  #-----  electrobergsolutions  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    FilmMaker:
      x-skaffolder-id: 5e2456330cfc6233ae9336f9
      x-skaffolder-id-db: 5e2456320cfc6233ae9336da
      x-skaffolder-id-entity: 5e2456330cfc6233ae9336f6
      x-skaffolder-url: /filmmaker
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5e2456330cfc6233ae9336f7
          x-skaffolder-type: String
          x-skaffolder-required: true
        surname:
          type: string
          x-skaffolder-id-attr: 5e2456330cfc6233ae9336f8
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
        - name
    User:
      x-skaffolder-id: 5e2456320cfc6233ae9336e4
      x-skaffolder-id-db: 5e2456320cfc6233ae9336da
      x-skaffolder-id-entity: 5e2456320cfc6233ae9336dc
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336df
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336e0
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336de
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336e2
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336e1
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336dd
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...